requirements.txt
easygui
python-dotenv
mysql-connector-python